FDG-6275 Initial refactor
