+++ src/layouts/explainer/sections/national-debt/growing-national-debt/growing-national-debt.module.scss
@@ -1,0 +1,14 @@
+
+
+
+.growingNationalDebt {
+  :global(.vis-callout) {
+    margin-top: calc(20% + 4rem);
+    margin-bottom: calc(20% + 4rem);
+  }
+}
+
+
+.postGraphAccordionContainer {
+  margin-top: 1rem;
+}
