+++ src/layouts/explainer/sections/national-debt/growing-national-debt/growing-national-debt.module.scss.d.ts
@@ -1,0 +1,3 @@
+// This file is automatically generated. Do not modify this file manually -- YOUR CHANGES WILL BE ERASED!
+export const growingNationalDebt: string;
+export const postGraphAccordionContainer: string;
