+++ src/layouts/explainer/sections/national-debt/national-debt.module.scss.d.ts
@@ -99,7 +99,6 @@ export const multichartLegend: string;
 export const debtLegend: string;
 export const lineChartContainer: string;
 export const postGraphContent: string;
-export const postGraphAccordionContainer: string;
 export const diveDeeperQuoteRight: string;
 export const diveDeeperQuote: string;
 export const diveDeeperQuoteLeft: string;
@@ -107,5 +106,4 @@ export const diveDeeperCitation: string;
 export const diveDeeperLink: string;
 export const debtAccordion: string;
 export const debtCeilingAccordion: string;
-export const growingNationalDebt: string;
 export const multichartWrapper: string;
