+++ src/pages/americas-finance-guide/afg-overview.module.scss.d.ts
@@ -105,7 +105,6 @@ export const multichartLegend: string;
 export const debtLegend: string;
 export const lineChartContainer: string;
 export const postGraphContent: string;
-export const postGraphAccordionContainer: string;
 export const diveDeeperQuoteRight: string;
 export const diveDeeperQuote: string;
 export const diveDeeperQuoteLeft: string;
@@ -113,7 +112,6 @@ export const diveDeeperCitation: string;
 export const diveDeeperLink: string;
 export const debtAccordion: string;
 export const debtCeilingAccordion: string;
-export const growingNationalDebt: string;
 export const multichartWrapper: string;
 export const deficitAccordion: string;
 export const mainContainer: string;
